change: 권한 및 비즈니스로직 테스트

+++ core/src/main/kotlin/com/wiyb/server/core/facade/CommunityFacade.kt
@@ -7,6 +7,8 @@ import com.wiyb.server.core.domain.community.PaginationQuery
 import com.wiyb.server.core.domain.community.PostDetailDto
 import com.wiyb.server.core.domain.community.UpdateCommentParam
 import com.wiyb.server.core.domain.community.UpdatePostParam
+import com.wiyb.server.core.domain.exception.CommonException
+import com.wiyb.server.core.domain.exception.ErrorCode
 import com.wiyb.server.core.service.CommunityService
 import com.wiyb.server.core.service.UserService
 import com.wiyb.server.storage.database.entity.common.dto.PaginationResultDto
@@ -21,8 +23,13 @@ class CommunityFacade(
     private val userService: UserService,
     private val communityService: CommunityService
 ) {
+    @Transactional
     fun getPostDetail(id: Long): PostDetailDto {
         val post = communityService.getPostDetail(id)
+
+        if (post.commentCount > 0) {
+            post.comments = communityService.getComments(id)
+        }
 
         communityService.increaseViewCount(id)
 
@@ -49,12 +56,26 @@ class CommunityFacade(
         val userProfile = userService.findUserProfileBySessionId(sessionId)
         val post = communityService.getPost(id)
 
+        if (!communityService.isAuthor(id, userProfile.id)) {
+            throw CommonException(ErrorCode.INSUFFICIENT_POST_AUTHORITY)
+        }
+
+        param.convert().let { post.update(it) }
         communityService.savePost(post)
 
         return PostDto.from(post, userProfile)
     }
 
-    fun deletePost(id: Long): Unit = communityService.deletePost(id)
+    fun deletePost(id: Long) {
+        val sessionId = SecurityContextHolder.getContext().authentication.name
+        val userProfile = userService.findUserProfileBySessionId(sessionId)
+
+        if (!communityService.isAuthor(id, userProfile.id)) {
+            throw CommonException(ErrorCode.INSUFFICIENT_POST_AUTHORITY)
+        }
+
+        communityService.deletePost(id)
+    }
 
     @Transactional
     fun createComment(
@@ -65,6 +86,10 @@ class CommunityFacade(
         val userProfile = userService.findUserProfileBySessionId(sessionId)
         val post = communityService.getPost(postId)
         val comment = param.to(userProfile, post)
+
+        if (param.replyTo != null && !communityService.unsafeCheckExistsComment(param.replyTo, postId)) {
+            throw CommonException(ErrorCode.COMMENT_REPLY_NOT_ALLOWED)
+        }
 
         communityService.saveComment(comment)
         communityService.increaseCommentCount(postId)
@@ -80,6 +105,11 @@ class CommunityFacade(
         val userProfile = userService.findUserProfileBySessionId(sessionId)
         val comment = communityService.getComment(path)
 
+        if (!communityService.isCommentAuthor(comment.id, userProfile.id)) {
+            throw CommonException(ErrorCode.INSUFFICIENT_POST_AUTHORITY)
+        }
+
+        param.convert().let { comment.update(it) }
         communityService.saveComment(comment)
 
         return CommentDto.from(comment, userProfile)
@@ -87,7 +117,15 @@ class CommunityFacade(
 
     @Transactional
     fun deleteComment(path: CommentIdPath) {
-        communityService.decreaseCommentCount(path.postId)
+        val sessionId = SecurityContextHolder.getContext().authentication.name
+        val userProfile = userService.findUserProfileBySessionId(sessionId)
+
+        if (!communityService.isCommentAuthor(path.commentId, userProfile.id)) {
+            throw CommonException(ErrorCode.INSUFFICIENT_POST_AUTHORITY)
+        }
+
+        // todo: 기획 따라서. 일단은 댓글 감소는 없는 걸로
+        // communityService.decreaseCommentCount(path.postId)
         communityService.deleteComment(path.commentId)
     }
 }
